fixed EditPage rerender + added item deletion on server side

+++ src/components/EditPageFilter/EditPageFilter.jsx
@@ -3,12 +3,6 @@ import s from './EditPageFilter.module.css';
 // import EditableProduct from './EditableProduct';
 import { v4 as uuidv4 } from 'uuid';
 import { getItems, getCategories, deleteItem } from '../../serverqueries';
-import {
-  getItems,
-  getCategories,
-  deleteItem,
-  // baseUrl,
-} from '../../serverqueries';
 
 // This holds a list of some fiction products
 // Some  have the same name but different cost and id
@@ -24,12 +18,11 @@ const initialProducts = [
 //   if (!stateProducts || !stateProducts.length) {
 //     return <h1>Loading</h1>;
 //   }
-function matchCategory(categoryId, categoriesListOfObj) {
-  const category = categoriesListOfObj.find(obj => obj.id === categoryId);
-  return category.name;
-}
 
-let tempVar = { products: [], categories: [] };
+function matchCategory(categoryId, categoriesListOfObj) {
+  const categoryObj = categoriesListOfObj.find(obj => obj.id === categoryId);
+  return categoryObj.name;
+}
 
 function EditPageFilter() {
   // the value of the search field
@@ -42,60 +35,42 @@ function EditPageFilter() {
   const [stateProducts, setStateProducts] = useState([]);
   const [stateCategories, setStateCategories] = useState([]);
 
-  function matchCategory(categoryId, categoriesListOfObj) {
-    const categoryObj = categoriesListOfObj.find(obj => obj.id === categoryId);
-    debugger;
 
-    return categoryObj.name;
-  }
   useEffect(() => {
     getItems().then(items => setStateProducts(items));
     getCategories().then(items => setStateCategories(items));
   }, []);
-  useEffect(() => {
-    deleteItem()
-      // async function dltItem(idToDelete, setStateProducts) {
-      //   await fetch(`${baseUrl}/product/${idToDelete}`, {
-      //     method: 'DELETE',
-      //     headers: { 'Content-Type': 'application/json' },
-      //     params: { id: idToDelete },
-      //     body: null,
-      //   })
-      // .then(res => {
-      //   getItems().then((res)=> {setState(res)})
-      // })
-      .then(() => getItems())
-      .then(items => setStateProducts(items))
-      .then(console.log(`Item deleted`));
-  }, []);
+
   const deleteItems = idToDelete => {
     setStateProducts(stateProducts.filter(item => item.id !== idToDelete));
-    console.log('delete from fetch');
+    deleteItem(idToDelete);
   };
-  let mappedStateToJsx = stateProducts.map(categoryObj => {
-    return (
-      <li key={categoryObj.id} className={s.product}>
-        <span className={s.id}>
-          {matchCategory(categoryObj.categoryId, stateCategories)}
-        </span>
-        <span className={s.name}>{categoryObj.name}</span>
+  
+  let mappedStateToJsx =[];
+  if (stateProducts.length > 0 && stateCategories.length > 0) {
+    mappedStateToJsx = stateProducts.map(categoryObj => {
+      return (
+        <li key={categoryObj.id} className={s.product}>
+          <span className={s.id}>
+            {matchCategory(categoryObj.categoryId, stateCategories)}
+          </span>
+          <span className={s.name}>{categoryObj.name}</span>
+          <button type="submit" className={s.button}>
+            Edit
+          </button>
+          <button
+            type="submit"
+            className={s.buttonDelete}
+            value={val}
+            onClick={() => deleteItems(categoryObj.id)}
+          >
+            Delete
+          </button>
+        </li>
+      );
+    })
+  };
 
-        <button type="submit" className={s.button}>
-          Edit
-        </button>
-        <button
-          type="submit"
-          className={s.buttonDelete}
-          value={val}
-          onClick={() => deleteItems(categoryObj.id)}
-        >
-          Delete
-        </button>
-      </li>
-    );
-  });
-
-  // let mappedStateToJsx = stateProducts.map(el => {});
   // the search result
   const [foundProducts, setFoundProducts] = useState(initialProducts);
